+++ jina/orchestrate/deployments/__init__.py
@@ -15,11 +15,12 @@ from typing import Dict, List, Optional, Set, Union
 from hubble.executor.helper import replace_secret_of_hub_uri
 from hubble.executor.hubio import HubIO
 
-from jina import __default_executor__, __default_host__, __docker_host__, helper
+from jina.constants import __default_executor__, __default_host__, __docker_host__
 from jina.enums import DeploymentRoleType, PodRoleType, PollingType
 from jina.helper import (
     CatchAllCleanupContextManager,
     parse_host_scheme,
+    random_port
 )
 from jina.orchestrate.deployments.install_requirements_helper import install_package_dependencies, _get_package_path_from_uses
 from jina.orchestrate.pods.factory import PodFactory
@@ -863,28 +864,28 @@ class Deployment(BaseDeployment):
 
                     elif shards == 1:
                         _args.port_monitoring = (
-                            helper.random_port()
+                            random_port()
                             if replica_id >= len(self.args.all_port_monitoring)
                             else self.args.all_port_monitoring[replica_id]
                         )
                         # if there are no shards/replicas, we dont need to distribute ports randomly
                         # we should rather use the pre assigned one
-                        _args.port = helper.random_port()
+                        _args.port = random_port()
                     elif shards > 1:
                         port_monitoring_index = (
                             replica_id + replicas * shard_id + 1
                         )  # the first index is for the head
                         _args.port_monitoring = (
-                            helper.random_port()
+                            random_port()
                             if port_monitoring_index >= len(self.args.all_port_monitoring)
                             else self.args.all_port_monitoring[
                                 port_monitoring_index
                             ]  # we skip the head port here
                         )
-                        _args.port = helper.random_port()
+                        _args.port = random_port()
                     else:
-                        _args.port = helper.random_port()
-                        _args.port_monitoring = helper.random_port()
+                        _args.port = random_port()
+                        _args.port_monitoring = random_port()
 
                 else:
                     _args.port = self.ext_repl_ports[replica_id]
