+++ jina/orchestrate/deployments/config/docker_compose.py
@@ -3,7 +3,7 @@ import os
 from argparse import Namespace
 from typing import Dict, List, Optional, Tuple, Union
 
-from jina import (
+from jina.constants import (
     __default_executor__,
     __default_grpc_gateway__,
     __default_http_gateway__,
