+++ jina/orchestrate/deployments/config/helper.py
@@ -4,12 +4,11 @@ from typing import Dict
 from hubble.executor.helper import is_valid_docker_uri, parse_hub_uri
 from hubble.executor.hubio import HubIO
 
-from jina import (
+from jina.constants import (
     __default_executor__,
     __default_grpc_gateway__,
     __default_http_gateway__,
     __default_websocket_gateway__,
-    __version__,
 )
 from jina.enums import PodRoleType
 
@@ -58,6 +57,7 @@ def get_base_executor_version():
     import requests
 
     try:
+        from jina import __version__
         url = 'https://registry.hub.docker.com/v2/repositories/jinaai/jina/tags'
         result: Dict = requests.get(url, params={'name': __version__}).json()
         if result.get('count', 0) > 0:
