+++ jina/orchestrate/deployments/config/k8s.py
@@ -2,7 +2,7 @@ import copy
 from argparse import Namespace
 from typing import Dict, List, Optional, Tuple, Union
 
-from jina import (
+from jina.constants import (
     __default_executor__,
     __default_grpc_gateway__,
     __default_http_gateway__,
