+++ jina/serve/executors/__init__.py
@@ -9,7 +9,7 @@ import warnings
 from types import SimpleNamespace
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Type, Union, overload
 
-from jina import __args_executor_init__, __cache_path__, __default_endpoint__
+from jina.constants import __args_executor_init__, __cache_path__, __default_endpoint__
 from jina.enums import BetterEnum
 from jina.helper import (
     ArgNamespace,
@@ -717,7 +717,7 @@ class BaseExecutor(JAMLCompatible, metaclass=ExecutorType):
         :param kwargs: other kwargs accepted by the Flow, full list can be found `here <https://docs.jina.ai/api/jina.orchestrate.flow.base/>`
 
         """
-        from jina import Flow
+        from jina.orchestrate.flow.base import Flow
 
         f = Flow(**kwargs).add(
             uses=cls,
@@ -768,7 +768,7 @@ class BaseExecutor(JAMLCompatible, metaclass=ExecutorType):
         :param uses_dynamic_batching: dictionary of parameters to overwrite from the default config's dynamic_batching field
         :param kwargs: other kwargs accepted by the Flow, full list can be found `here <https://docs.jina.ai/api/jina.orchestrate.flow.base/>`
         """
-        from jina import Flow
+        from jina.orchestrate.flow.base import Flow
 
         Flow(**kwargs).add(
             uses=uses,
@@ -811,7 +811,7 @@ class BaseExecutor(JAMLCompatible, metaclass=ExecutorType):
         :param uses_dynamic_batching: dictionary of parameters to overwrite from the default config's requests field
         :param kwargs: other kwargs accepted by the Flow, full list can be found `here <https://docs.jina.ai/api/jina.orchestrate.flow.base/>`
         """
-        from jina import Flow
+        from jina.orchestrate.flow.base import Flow
 
         f = Flow(**kwargs).add(
             uses=uses,
